+++ TODO.md
@@ -1,0 +1,11 @@
+TODO
+===
+
+- [x] Add asciinema
+- [x] Add [ora](https://www.npmjs.com/package/ora) spinner
+- [x] Add [request-promise](https://www.npmjs.com/package/request-promise)
+- [ ] Update text errors
+- [ ] Handle Ctrl+u and backspace in prompt password
+- [ ] Handle multi-link
+- [ ] Add in AUR repository
+- [ ] Update token after expiration
