+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "rdcli",
-  "version": "1.0.18",
+  "version": "1.1.0",
   "description": "Command Line Interface for real-debrid.com",
   "homepage": "https://github.com/jcherqui/rdcli",
   "repository": {
@@ -21,7 +21,7 @@
   "author": "jcherqui",
   "license": "MIT",
   "dependencies": {
-    "babel-polyfill": "^6.8.0",
+    "babel-polyfill": "^6.9.1",
     "chai": "^3.5.0",
     "chalk": "^1.1.3",
     "co": "^4.6.0",
@@ -32,23 +32,24 @@
     "fs": "^0.0.2",
     "humanize": "^0.0.9",
     "ora": "^0.2.3",
-    "request": "^2.69.0",
+    "request": "^2.72.0",
     "request-progress": "^2.0.1",
     "url": "^0.11.0"
   },
   "devDependencies": {
-    "babel-cli": "^6.8.0",
-    "babel-core": "^6.8.0",
-    "babel-preset-es2015": "^6.6.0",
-    "body-parser": "^1.15.1",
+    "babel-cli": "^6.10.1",
+    "babel-core": "^6.10.4",
+    "babel-preset-es2015": "^6.9.0",
+    "body-parser": "^1.15.2",
     "co-mocha": "^1.1.2",
-    "eslint": "^2.9.0",
+    "eslint": "^2.13.1",
     "eslint-config-airbnb": "^9.0.1",
-    "eslint-plugin-import": "^1.7.0",
-    "eslint-plugin-jsx-a11y": "^1.0.4",
-    "eslint-plugin-react": "^5.0.1",
-    "json-server": "^0.8.10",
-    "mocha": "^2.4.5"
+    "eslint-plugin-import": "^1.9.2",
+    "eslint-plugin-jsx-a11y": "^1.5.3",
+    "eslint-plugin-react": "^5.2.2",
+    "json-server": "^0.8.14",
+    "mocha": "^2.5.3",
+    "request-promise": "^3.0.0"
   },
   "keywords": [
     "real-debrid",
