+++ src/rdcli.js
@@ -1,11 +1,15 @@
 #!/usr/bin/env node
 
-import RealDebrid from './real-debrid';
+import { download, waitDuringScan } from './download';
+import { convertTorrent, convertMagnet } from './torrent';
+import unrestrict from './unrestrict';
+import getToken from './connect';
 import pjson from '../package.json';
 import program from 'commander';
 import prompt from 'co-prompt';
 import ora from 'ora';
 import co from 'co';
+import chalk from 'chalk';
 import fs from 'fs';
 
 program
@@ -13,45 +17,43 @@ program
     .description('Download links, magnets and torrent files.')
     .usage('<url|magnet|torrent>')
     .action((arg) => co(function*action() {
-        const api = new RealDebrid();
-
-        const connect = function*connect() {
+        try {
             const username = yield prompt('Username: ');
             const password = yield prompt.password('Password: ');
-            yield api.connect(username, password);
-        };
+            const token = yield getToken(username, password);
 
-        let link;
-        if (arg.match(/^(https?:\/\/)([\da-z\.-]+)\.([a-z\.]{2,6})([\/\w\? \.-]*)*\/?$/)) {
-            yield connect();
-            link = arg;
-        } else if (arg.match(/^magnet:\?xt=urn:[a-z0-9]+:[a-z0-9]{20,50}/i)) {
-            yield connect();
-            link = yield api.convertTorrent(arg);
-        } else if (fs.existsSync(arg)) {
-            yield connect();
-            link = yield api.convertTorrent(arg);
-        } else {
-            console.log('Usage: rdcli <url|magnet|torrent>');
+            let link;
+            if (arg.match(/^(https?:\/\/)([\da-z\.-]+)\.([a-z\.]{2,6})([\/\w\? \.-]*)*\/?$/)) {
+                link = arg;
+            } else if (arg.match(/^magnet:\?xt=urn:[a-z0-9]+:[a-z0-9]{20,50}/i)) {
+                link = yield convertMagnet(arg, token);
+            } else if (fs.existsSync(arg)) {
+                link = yield convertTorrent(arg, token);
+            } else {
+                console.log('Usage: rdcli <url|magnet|torrent>');
+                process.exit();
+            }
+
+            const unrestrictLink = yield unrestrict(link, token);
+            console.log(`Start download : ${link}`);
+            yield waitDuringScan(link);
+
+            const spinner = ora('Download: 0.0% Speed: 0Mbps').start();
+            download(unrestrictLink, (res) => {
+                if (res.percent) {
+                    spinner.text = `Download: ${res.percent}% Speed: ${res.mbps} ${res.bytesWriting}/${res.totalSize} Remaining: ${res.remaining}sec`; // eslint-disable-line max-len
+                } else if (res === 'end') {
+                    spinner.stop();
+                    console.log('File downloaded.');
+                } else {
+                    spinner.stop();
+                    console.log(`Error: ${res}`);
+                }
+            });
+        } catch (e) {
+            console.error(`\n${chalk.red(e)}`);
             process.exit();
         }
-
-        const unrestrictLink = yield api.unrestrictLink(link);
-        console.log(`Start download : ${link}`);
-        yield api.waitDuringScan(link);
-
-        const spinner = ora('Download: 0.0% Speed: 0Mbps').start();
-        yield api.download(unrestrictLink, (res) => {
-            if (res.percent) {
-                spinner.text = `Download: ${res.percent}% Speed: ${res.mbps} ${res.bytesWriting}/${res.totalSize} Remaining: ${res.remaining}sec`; // eslint-disable-line max-len
-            } else if (res === 'end') {
-                spinner.stop();
-                console.log('File downloaded.');
-            } else {
-                spinner.stop();
-                console.log(`Error: ${res}`);
-            }
-        });
     }))
     .parse(process.argv);
